requirements.txt
python-dotenv==0.12.0
Django==2.2.11
django-cacheops==4.2
djangorestframework==3.11.0
djongo==1.3.1
funcy==1.14
pymongo==3.10.1

